requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.3
